worker threads started -> game intialization and loop moved to worker, refactor of emitters logic and state broadcast at end of Game loop
