+++ src/sockets/handlers/listeners.ts
@@ -1,7 +1,8 @@
-import { sendStartSignal } from "./emitters";
+import { broadCastState, sendStartSignal } from "./emitters";
 import intializeGame from "../../engine/functions/intializeGame";
 import { sv_games } from "../../utils/sv_games";
 import { Socket, DefaultEventsMap } from "socket.io";
+import { Worker } from "worker_threads";
 
 export default function attachListeners(
   socket: Socket<DefaultEventsMap, DefaultEventsMap, DefaultEventsMap, any>
@@ -13,11 +14,25 @@ export default function attachListeners(
   });
 
   socket.on("sv_start", (id: number) => {
-    console.log(`launching Game ${id}`);
-    const nwGame = intializeGame(id);
-    sv_games[id] = nwGame;
+    //console.log(`launching Game ${id}`);
+    // const nwGame = intializeGame(id);
+    // sv_games[id] = nwGame;
     console.log("Sending client start message...");
+
+    // assign game instnace to assigned worker thread for this socket
+
     sendStartSignal();
+
+    const worker = new Worker("./dist/workers/game-worker.js", {
+      workerData: id,
+    });
+    worker.on("message", (msg) => {
+      if (msg.start) console.log(msg.start);
+      else if (msg.state) {
+        //console.log(msg.state);
+        broadCastState(msg.state);
+      }
+    });
   });
 
   socket.on("disconnect", () => {
